exp fixes

+++ common/src/main/java/net/mehvahdjukaar/polytone/common/expressions/proxies/GlobalProxy.java
@@ -17,7 +17,7 @@ public class GlobalProxy {
         return Minecraft.getInstance().level;
     }
 
-    public long time(){
+    public double time(){
         return delegate().getGameTime();
     }
